minor bug fix

+++ src/main/java/lk/apiit/eea1/online_crafts_store/ReviewsAndRatings/Repository/RateCraftCreatorRepository.java
@@ -14,5 +14,5 @@ public interface RateCraftCreatorRepository extends JpaRepository<RateCraftCreat
 
     int countAllByCraftCreator(CraftCreator craftCreator);
 
-    int countAllByRatingAndCraftCreator(int rating,CraftCreator craftCreator);
+    int countAllByRatingAndCraftCreator(float rating,CraftCreator craftCreator);
 }
